+++ spec/oanda_api/streaming/request_spec.rb
@@ -3,6 +3,7 @@ require 'uri'
 require 'webmock/rspec'
 
 describe "OandaAPI::Streaming::Request" do
+
   let(:streaming_request) {
     OandaAPI::Streaming::Request.new(uri: "https://a.url.com",
                                      query: { account: 1234, instruments: %w[AUD_CAD AUD_CHF] },
@@ -60,7 +61,10 @@ describe "OandaAPI::Streaming::Request" do
       END
       ids = []
       stub_request(:any, /\.com/).to_return(body: events_json, status: 200)
-      streaming_request.stream { |resource| ids << resource.id }
+
+      streaming_request.stream do |resource|
+        ids << resource.id
+      end
       expect(ids).to contain_exactly(1, 2)
     end
 
@@ -166,6 +170,30 @@ describe "OandaAPI::Streaming::Request" do
           signaller.stop!
         end
         expect(heartbeats).to eq 1
+      end
+    end
+
+    context "when the stream contains multiple undelimited objects" do
+      events_json = <<-END
+        {"tick":{"bid": 1}}{"tick":{"bid": 2}}\r\n{"tick":{"bid": 3}}
+      END
+
+      it "yields all of the objects" do
+        case
+        when gem_installed?(:Yajl)
+          OandaAPI::Streaming::JsonParser.use(:yajl)
+        when gem_installed?(:Gson)
+          OandaAPI::Streaming::JsonParser.use(:gson)
+        else
+          OandaAPI::Streaming::JsonParser.use(:generic)
+        end
+
+        stub_request(:any, /\.com/).to_return(body: events_json, status: 200)
+        ticks = []
+        streaming_request.stream do |resource|
+          ticks << resource.bid
+        end
+        expect(ticks).to contain_exactly(1, 2, 3)
       end
     end
   end